Add Auth and Noauth filters, Add Login and register with profile  components
The project is at a point of basic functionality as follows, a user can log in and add products of the system and view the store with products listed with a category tabs menu. A user can view a single product as well as editing it, they can also see their order only after placing it. Finally the user has the ability to edit their profile and they can logout as well.